feat(json-rpc): move to snake-case
